+++ test/print.cpp
@@ -23,34 +23,31 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-/*!
-  \file kitty.hpp
-  \brief Main header for kitty
+#include <gtest/gtest.h>
 
-  \author Mathias Soeken
-*/
+#include <sstream>
 
-#pragma once
+#include <kitty/kitty.hpp>
 
-#include "static_truth_table.hpp"
-#include "dynamic_truth_table.hpp"
+#include "utility.hpp"
 
-#include "algorithm.hpp"
-#include "bit_operations.hpp"
-#include "canonization.hpp"
-#include "constructors.hpp"
-#include "hash.hpp"
-#include "isop.hpp"
-#include "operations.hpp"
-#include "operators.hpp"
-#include "print.hpp"
+using namespace kitty;
 
-/*
-         /\___/\
-        (  o o  )
-        /   *   \
-        \__\_/__/
-          /   \
-         / ___ \
-         \/___\/
-*/
+class PrintTest : public kitty::testing::Test
+{
+protected:
+  template<typename TT>
+  std::string to_binary( const TT& tt ) const
+  {
+    std::stringstream ss;
+    print_binary( tt, ss );
+    return ss.str();
+  }
+};
+
+TEST_F( PrintTest, print_binary )
+{
+  EXPECT_EQ( to_binary( from_hex<2>( "8" ) ), "1000" );
+  EXPECT_EQ( to_binary( from_hex<3>( "e8" ) ), "11101000" );
+  EXPECT_EQ( to_binary( from_hex<7>( "fffefee8fee8e880fee8e880e8808000" ) ), "11111111111111101111111011101000111111101110100011101000100000001111111011101000111010001000000011101000100000001000000000000000" );
+}